+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-07-26 17:27:34
-// Commit: docs(core): update readme and API contracts for core
+// Updated at: 2026-07-27 11:16:07
+// Commit: feat(core): add bi-directional event bus handler for core
 
 class coreComponentState {
-  final String commitRef = '3160';
-  final String lastMessage = 'docs(core): update readme and API contracts for core';
-  final DateTime timestamp = DateTime.parse('2026-07-26T17:27:34+05:00');
+  final String commitRef = '3170';
+  final String lastMessage = 'feat(core): add bi-directional event bus handler for core';
+  final DateTime timestamp = DateTime.parse('2026-07-27T11:16:07+05:00');
 }